Consolidate fallback tests into test_python_ops_fallback.py
